requirements.txt
Flask>=2.0.0
openpyxl>=3.0.0
pandas==1.4.2
requests==2.32.5
requests-file>=1.5.0
dashscope>=1.14.0
python-dotenv>=0.19.0
alibabacloud_bailian20231229
alibabacloud_tea_openapi
beautifulsoup4>=4.9.0
